commit for adding test validation on search user exist AND adding comments on PlaylistService and all Repository

+++ core/src/main/java/userCase/PlaylistService.java
@@ -10,7 +10,6 @@ import repository.PlaylistRepository;
 import repository.UserRepository;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 public class PlaylistService {
 
@@ -28,93 +27,152 @@ public class PlaylistService {
         return this.playlistRepository.getPlaylistByUser(user);
     }
 
+    //Use case create a playlist
     public Response createPlaylist(PlaylistDTO playlistDTO, UserRepository userRepository, String email){
+
+        //Verify errors about field are all filled and correct
         List<ValidationError> validationErrors = validateCreate(playlistDTO);
         if (!validationErrors.isEmpty()) {
-            //throw new ValidationErrorException(validationErrors);
             Response responseError = new Response(false, "Error of validation on creation of playlist", validationErrors);
             return responseError;
         }
 
+        //Verify errors about searching if user exist (playlist is link to this user)
+        List<ValidationError> validationErrorsForSearchUser = validateSearchUser(userRepository, email);
+        if (!validationErrorsForSearchUser.isEmpty()) {
+            Response responseError = new Response(false, "Error of search user", validationErrorsForSearchUser);
+            return responseError;
+        }
+
+        //Getting the user by his email
         User user = userRepository.getUserByEmail(email);
 
+        //Creation of user and delete it before add it (for no double)
         Playlist newPlaylist = new Playlist(playlistDTO.getNamePlaylist(), user);
         this.playlistRepository.deletePlaylist(newPlaylist);
         this.playlistRepository.addPlaylist(newPlaylist);
 
+        //Response validation
         Response responseValidate = new Response(true, "Validation on creation of playlist", validationErrors);
         return responseValidate;
     }
 
+    //Use case add song to a playlist
     public Response addSongToPlaylist(Song song, String email, String namePlaylist, UserRepository userRepository){
 
+        //Verify errors about searching if playlist exist
         List<ValidationError> validationErrors = validateSearchPlaylist(namePlaylist);
         if (!validationErrors.isEmpty()) {
-            //throw new ValidationErrorException(validationErrors);
             Response responseError = new Response(false, "Error of search the playlist", validationErrors);
             return responseError;
         }
 
-        User user = userRepository.getUserByEmail(email);
-        Playlist p = this.playlistRepository.getOnePlaylistByName(namePlaylist, user);
-        p.addSongToPlaylist(song);
-        //this.playlistRepository.getPlaylistById(idPlaylist).addSongToPlaylist(song);
+        //Verify errors about searching if user exist (playlist is link to this user)
+        List<ValidationError> validationErrorsForSearchUser = validateSearchUser(userRepository, email);
+        if (!validationErrorsForSearchUser.isEmpty()) {
+            Response responseError = new Response(false, "Error of search user", validationErrorsForSearchUser);
+            return responseError;
+        }
 
+        //Getting the user by his email
+        User user = userRepository.getUserByEmail(email);
+
+        //Getting the playlist by the name of playlist associate to user
+        Playlist p = this.playlistRepository.getOnePlaylistByName(namePlaylist, user);
+
+        //Add song to the playlist
+        p.addSongToPlaylist(song);
+
+        //Response validation
         Response responseValidate = new Response(true, "Validation on add song to playlist", validationErrors);
         return responseValidate;
     }
 
+    //Use case delete song from a playlist
     public Response deleteSongFromPlaylist(Song song, String email, String namePlaylist, UserRepository userRepository){
-        List<ValidationError> validationErrorsForSearchPlaylist = validateSearchPlaylist(email);
+
+        //Verify errors about searching if playlist exist
+        List<ValidationError> validationErrorsForSearchPlaylist = validateSearchPlaylist(namePlaylist);
         if (!validationErrorsForSearchPlaylist.isEmpty()) {
-            //throw new ValidationErrorException(validationErrors);
             Response responseError = new Response(false, "Error of search the playlist", validationErrorsForSearchPlaylist);
             return responseError;
         }
 
+        //Verify errors about searching if song exist in playlist
         List<ValidationError> validationErrorsForSongInPlaylist = validateSongIsInPlaylist(namePlaylist, song);
         if (!validationErrorsForSongInPlaylist.isEmpty()) {
-            //throw new ValidationErrorException(validationErrors);
             Response responseError = new Response(false, "Error of search song in the playlist", validationErrorsForSongInPlaylist);
             return responseError;
         }
 
+        //Verify errors about searching if user exist (playlist is link to this user)
+        List<ValidationError> validationErrorsForSearchUser = validateSearchUser(userRepository, email);
+        if (!validationErrorsForSearchUser.isEmpty()) {
+            Response responseError = new Response(false, "Error of search user", validationErrorsForSearchUser);
+            return responseError;
+        }
+
+        //Getting the playlist by the name of playlist and the user
+        //AND THEN
+        //Delete the song from the playlist
         this.playlistRepository.getPlaylistByName(namePlaylist).deleteSongFromPlaylist(song);
 
+        //Response validation
         Response responseValidate = new Response(true, "Validation on delete song from playlist", validationErrorsForSongInPlaylist);
         return responseValidate;
     }
 
+    //Use case delete a playlist
     public Response deletePlaylist(String namePlaylist){
+
+        //Verify errors about searching if playlist exist
         List<ValidationError> validationErrors = validateSearchPlaylist(namePlaylist);
         if (!validationErrors.isEmpty()) {
-            //throw new ValidationErrorException(validationErrors);
             Response responseError = new Response(false, "Error of search the playlist", validationErrors);
             return responseError;
         }
 
+        //Getting the playlist by the name of playlist
+        //AND THEN
+        //Delete the playlist
         this.playlistRepository.deletePlaylist(this.playlistRepository.getPlaylistByName(namePlaylist));
 
+        //Response validation
         Response responseValidate = new Response(true, "Validation on delete playlist", validationErrors);
         return responseValidate;
     }
 
+    //Use case share a playlist
     public Response sharePlaylist(String email, UserRepository userRepository, String namePlaylist){
+
+        //Verify errors about searching if user exist (playlist is link to this user)
         List<ValidationError> validationErrors = validateSearchUser(userRepository, email);
         if (!validationErrors.isEmpty()) {
-            //throw new ValidationErrorException(validationErrors);
             Response responseError = new Response(false, "Error of search the user", validationErrors);
             return responseError;
         }
 
+        //Verify errors about searching if playlist exist
+        List<ValidationError> validationErrorsSearchPlaylist = validateSearchPlaylist(namePlaylist);
+        if (!validationErrorsSearchPlaylist.isEmpty()) {
+            Response responseError = new Response(false, "Error of search the playlist", validationErrorsSearchPlaylist);
+            return responseError;
+        }
+
+        //Getting the user by his email
         User user = userRepository.getUserByEmail(email);
 
+        //Getting the playlist by the name of playlist
+        //AND THEN
+        //Add a user to this playlist
         this.playlistRepository.getPlaylistByName(namePlaylist).addUserToPlaylist(user);
 
+        //Response validation
         Response responseValidate = new Response(true, "Validation on add user to playlist", validationErrors);
         return responseValidate;
     }
 
+    //ALL VALIDATION METHODS BELOW
     private List<ValidationError> validateCreate(PlaylistDTO playlistDTO) {
         List<ValidationError> listError = new ArrayList<>();
